+++ auparse/mk_audit_msg_types.go
@@ -15,6 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//go:build ignore
 // +build ignore
 
 package main
@@ -157,13 +158,11 @@ func GetAuditMessageType(name string) (AuditMessageType, error) {
 
 var tmpl = template.Must(template.New("message_types").Parse(fileTemplate))
 
-var (
-	headers = []string{
-		`https://raw.githubusercontent.com/torvalds/linux/v4.16/include/uapi/linux/audit.h`,
-		`https://raw.githubusercontent.com/linux-audit/audit-userspace/4d933301b1835cafa08b9e9ef705c8fb6c96cb62/lib/libaudit.h`,
-		`https://raw.githubusercontent.com/linux-audit/audit-userspace/4d933301b1835cafa08b9e9ef705c8fb6c96cb62/lib/msg_typetab.h`,
-	}
-)
+var headers = []string{
+	`https://raw.githubusercontent.com/torvalds/linux/v5.16/include/uapi/linux/audit.h`,
+	`https://raw.githubusercontent.com/linux-audit/audit-userspace/v3.0.7/lib/libaudit.h`,
+	`https://raw.githubusercontent.com/linux-audit/audit-userspace/v3.0.7/lib/msg_typetab.h`,
+}
 
 func DownloadFile(url, destinationDir string) (string, error) {
 	resp, err := http.Get(url)